moved query to the backend
