+++ .gitignore
@@ -101,3 +101,5 @@ typings/
 
 # TernJS port file
 .tern-port
+
+dist/**/*
