+++ backend/app/SeriesProvider.ts
@@ -1,6 +1,6 @@
 import { sample } from 'lodash';
 import * as math from 'mathjs';
-import { RANDOM_WORDS } from '../config';
+import { RANDOM_WORDS } from '../../shared/config';
 import { InfinityQuery, dataPoint, DataOverride } from '../../shared/types';
 
 const getStepFromRange = (startTime: number, endTime: number): number => {
