+++ shared/config.ts
