+++ src/module.ts
@@ -1,10 +1,10 @@
 import { DataSourcePlugin } from '@grafana/data';
-import { Datasource } from './datasource';
+import { DataSource } from './datasource';
 import { InfinityConfigEditor } from './config.editor';
 import { QueryEditor } from './query.editor';
 import { InfinityAnnotationCtrl } from './annotations.editor';
 
-export const plugin = new DataSourcePlugin(Datasource)
+export const plugin = new DataSourcePlugin(DataSource)
   .setConfigEditor(InfinityConfigEditor)
   .setQueryEditor(QueryEditor)
   .setAnnotationQueryCtrl(InfinityAnnotationCtrl);
